CRUD de MedVet funcionando. Busca CRMV separada de busca por ID. DeleteMedVet funciona por crmv. É preciso pensar em fechamento de crmv sem deletar usuario/medvet nas próximas etapas (novos métodos, considerando que medvets podem ter crmv secundários).

+++ src/main/java/br/edu/ifrs/miguelzk/application/dto/MedVetRequestDTO.java
@@ -18,6 +18,6 @@ public class MedVetRequestDTO {
     private String nomeCompleto;
 //    private DateJavaType dataNascimento;
 //    private DateJavaType dataCadastro;
-    private int crmv;
+    private Long crmv;
     private String especialidade;
 }
